changed envdata model to use double(4,1) to accomodate 100s

+++ models/hvac/EnvData.js
@@ -1,11 +1,11 @@
 module.exports = function(sequelize, DataTypes) {
 	return sequelize.define('EnvData', {
 		temperature: {
-			type: DataTypes.DOUBLE(3,1)
+			type: DataTypes.DOUBLE(4,1)
 			,allowNull: true
 		}
 		,humidity: {
-			type: DataTypes.DOUBLE(3,1)
+			type: DataTypes.DOUBLE(4,1)
 			,allowNull: true
 		}
 	});
